+++ examples/complex_list/lib/main.dart
@@ -10,9 +10,8 @@ void main() {
   Bloc.observer = const SimpleBlocObserver();
   runApp(BlocEaseStateWidgetsProvider(
       initialStateBuilder: () => const Placeholder(),
-      loadingStateBuilder: ([progress]) =>
-          const Center(child: CircularProgressIndicator()),
-      failureStateBuilder: ([exceptionObject, failureMessage]) =>
+      loadingStateBuilder: ([progress]) => const Center(child: CircularProgressIndicator()),
+      failureStateBuilder: ([failureMessage, exception, retryCallback]) =>
           Center(child: Text(failureMessage ?? 'Oops something went wrong!')),
       child: App(repository: Repository())));
 }